Change the prompt to a pseudoelement
Problem:
It was possible to remove a prompt using backspace, because it was
a nested span element, which could be targeted with DOM api methods
f.e. textContent which is a concatenated string of all children nodes
(including the prompt in this case).
Solution:
To prevent this problem I moved a prompt to css ::before pseudo-element.