Extract terminal's state into a separate file
Problem:
1. Im not using here any module system, neither CommonJS, ES6+ modules, webpack.
The only option I have here is to use a standard, older approach of keeping track
of all dependencies. To make things logical I chose to move the whole terminal object
to the new file and put it at the top of my script links
so I can rest assured that not error arises when trying to
access it from the other files.
2. If I'd want to write some "subprograms" to run on this terminal I'd have hard
time capturing input, as far as I know there is no method that
would "stop" execution of the script so I can wait for the user input (Other than 'prompt' from Browser API). Obviously
infinite loop by itself wouldnt be a good choice because my whole webpage
would stop and the web browser would crash eventually.
Solution:
The solution addresses the 2nd problem (the 1st problem is pretty self-explanatory).
To solve the issue and
  a. Still use the same input
  b. Ignore actual commands like /help, rps
I decided to make a special property in terminal object (called eventAwait)
that will keep a reference to the function that is currently using input
and once the user triggers keydown event and presses enter (sends a command)
the function kept in eventAwait will be called. So it's somewhat similiar to
eventEmitters.