Implement a main function for RPS game
The function is a wrapper for actual rps function with an infinite loop in it.
I decided to make it a generator function because Im using an infinite while loop
inside of it so it either would crash the web browser or If i'd change the
condition it would execute instantly looping all 5 times in a split of second
so the user makes a choice only once and the answers is the same for all the rounds.